Used Long instead of int for Integer values
git-svn-id: https://svn.apache.org/repos/asf/directory/shared/branches/shared-schema@825793 13f79535-47bb-0310-9956-ffa450edef68

+++ ldap/src/main/java/org/apache/directory/shared/ldap/schema/comparators/IntegerOrderingComparator.java
@@ -80,8 +80,8 @@ public class IntegerOrderingComparator extends LdapComparator<String>
         // Both object must be stored as String for boolean
         // values. If this is not the case, we have a pb...
         // However, the method will then throw a ClassCastException
-        int b1 = Integer.parseInt( backendValue );
-        int b2 = Integer.parseInt( assertValue );
+        long b1 = Long.parseLong( backendValue );
+        long b2 = Long.parseLong( assertValue );
         
         if ( b1 == b2 )
         {
